+++ History.md
@@ -6,3 +6,5 @@
 * Add(): ZMQ PipeLine 模型: PipeLineVentilator, PipeLineSink, PipeLineBroker；应用场景并行处理大数据
 * Add(): ZMQ 多个服务器: MClientMode, MServerMode, MServerMode,可开启多个服务器；应用场景代理与反代理
 * Add(): ZMQ 线程间的信号传输：MTRelay
+* Add(): ZMQ 节点协调：SyncSub, SyncPub;发布者启动后会先等待所有订阅者进行连接，也就是节点协调。
+			每个订阅者会使用另一个套接字来告知发布者自己已就绪